requirements.txt
bertopic
textract
pdfminer.six
python-docx
dearpygui